dotbash_hist: move history handling
I'm trying to get closer to the ideal world where history is gracefully
shared across all instances of bash. To this end each bash saves its own
history file but we combine it into a search able history file and use
hh when we have it.

Also we trap on shell exit to flush history and remove the history file
to avoid too much pollution of the file space. We save the history files
in ~/.cache as well.