dotbash_keys: clean-up handling
The old code was too complex. As we are always using keychain just
ignore it if it doesn't exist. We should only start agents on login
shells if we need to - otherwise inherit everything. When inheriting
ensure we kill all agents. Otherwise clear keys on login.

I think I'm mostly moving to fish as a login shell now but having bash
shells work the same makes sense.